changes data types, and fixed isse with nullTime

+++ lib/profile/basal/helper.h
@@ -1,10 +1,12 @@
 #ifndef __HELPER_H_
 #define __HELPER_H_
 
+#include <inttypes.h>
 #include <algorithm>
 #include <vector>
 #include <string>
-#include <iostream>
+#include <ctime>
+
 
 #include "schedule.h"
 
@@ -14,9 +16,9 @@ class Helper
       Helper();
       ~Helper();
       std::vector<Schedule*> schedule;
-      int Minutes(time_t &givenTime);
+      int32_t Minutes(time_t &givenTime);
       time_t StringToTime(const char * now);
-      void AddToSchedule(const char * start, double rate, int minutes);
+      void AddToSchedule(const char * start, float rate, int32_t minutes);
       std::vector<Schedule*> SortSchedule(std::vector<Schedule> &schedule);
       std::vector<Schedule*> GetSchedule();
 };
